requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
requests>=2.31.0
razorpay>=1.4.2
